create view to show cart details

+++ GeekShopping.Web/Models/CartHeaderViewModel.cs
@@ -5,6 +5,6 @@
         public Int64 Id { get; set; }
         public string UserId { get; set; }
         public string CouponCode { get; set; }
-        public double PurchaseAmount { get; set; }
+        public decimal PurchaseAmount { get; set; }
     }
 }
